% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Device.R
\docType{class}
\name{Device-class}
\alias{Device-class}
\title{An S4 class to represent Device objects}
\value{
An object of a class Device
}
\description{
\code{new("Device")} will instatiate an object of class Device which holds
attributes and methods for useful for accessing the \href{https://open.fda.gov/apis/}{openFDA API}
}
\section{Slots}{

\describe{
\item{\code{ftTab}}{data.frame object of 510(k)}

\item{\code{classTab}}{data.frame object of Classification}

\item{\code{recTab}}{data.frame object of Recall Enforcement Reports}

\item{\code{aeTab}}{data.frame object of Adverse Events.}

\item{\code{pmaTab}}{data.frame object of Pre-market Approval}

\item{\code{rcTab}}{data.frame object of Recalls}

\item{\code{rlTab}}{data.frame object of Registrations and Listings}

\item{\code{udiTab}}{data.frame object of Unique Device Identifier}
}}

\examples{
drug1 <- new("Device", apikey="Sd8Jfad834FDS")

\dontrun{
  drugnorun <- new("Device")
}
}
\seealso{
\url{https://open.fda.gov/apis/drug/ndc/} for the actual
API description from openFDA. \code{\link{jsonlite}} for accessing API.
}
