% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFDA.R
\docType{methods}
\name{extractFDA}
\alias{extractFDA}
\alias{extractFDA,character,ANY-method}
\alias{extractFDA,FDA-method}
\alias{extractFDA,Other-method}
\title{An S4 method to extract information from the open FDA API}
\usage{
extractFDA(Obj, endpointtype, query, ...)

\S4method{extractFDA}{FDA}(Obj, endpointtype, query, ...)

\S4method{extractFDA}{Other}(Obj, endpointtype, query, ...)
}
\arguments{
\item{Obj}{An Obj of class FDA or child thereof.}

\item{endpointtype}{The endpoint after the class endpoint. For example,
If the \code{Obj} is of type indicating the endpoint for the specific class.
For example, for the class \code{Drug}, the endpointtype may be \code{"event"}.}

\item{query}{string of query to access FDA API. For example,
If we want to do \code{"https://api.fda.gov/drug/event.json?search=receivedate:[20040101+TO+20081231]&limit=1"},
we would do \code{extractFDA(drugObj, endpointtype="event", query="receivedate:[20040101+TO+20081231]&limit=1"}}

\item{\dots}{Passed down to searchlist within the function}
}
\value{
An object of a class data.frame
}
\description{
\code{extractFDA(Obj, endpointtype, query)} for classes other than \code{Other}
will extract tables based on the \code{query}. The \code{Other} class only has one
API endpoint therefore doesn't require endpointtype (i.e. \code{extractFDA(Obj, endpointtype, query)})
}
\examples{
drug1 <- new("Drug", apikey="Y3CgaZj67AotB7b4XLhzHJTY7oBWKUC3u1fYulw8")
drug1tab <- extractFDA(drug1, "event", "receivedate:[20040101+TO+20081231]&limit=1")

\dontrun{
  drugnorun <- extractFDA("Food")
  drugnorun <- extractFDA(drug1, "random") # not a valid access endpoint
  drugnorun <- extractFDA(drug1, "event", "randomquery") # not a valid query
}
}
\seealso{
\url{https://open.fda.gov/apis/drug/ndc/} for the actual
API description from openFDA. \code{\link{jsonlite}} for accessing API.
}
