% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-FDA.R
\docType{class}
\name{FDA-class}
\alias{FDA-class}
\alias{FDA}
\title{An S4 class to represent FDA objects}
\value{
An object of a class FDA.
}
\description{
\code{new("FDA")} will instatiate an object of class FDA which holds
attributes and methods for useful for accessing the \href{https://open.fda.gov/apis/}{openFDA API}
}
\section{Slots}{

\describe{
\item{\code{url}}{Base url. Default = \code{"https://api.fda.gov/"}.}

\item{\code{apikey}}{String of API key from  \href{https://open.fda.gov/apis/}{openFDA API}}

\item{\code{endpointfront}}{String of API access point. This will vary by different classess. 
For example, the \code{Drug} class will have \code{"drug"} as the \code{endpointfront} value.}

\item{\code{endpointback}}{String of API access point. This will vary by different classess. 
For example, the \code{Drug} class will have a variety of options based on the 
 \code{\link{listSF}}. Specifically \code{"event"} may be an option.}

\item{\code{searchfield}}{String of search field provided by open FDA. 
This will vary by different classess.}
}}

\examples{
drug1 <- new("FDA", apikey="Sd8Jfad834FDS")

\dontrun{
  drugnorun <- new("FDA")
}
}
\seealso{
\url{https://open.fda.gov/apis/drug/ndc/} for the actual
API description from openFDA. \code{\link{jsonlite}} for accessing API.
}
