% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDAopenR-package.R
\docType{package}
\name{FDAopenR-pkg}
\alias{FDAopenR-pkg}
\alias{FDAopenR-pkg-package}
\title{\pkg{FDAopenR}: R package for accessing and manipulating openFDA API.}
\description{
INSERT HERE DESCRIPTION OF THE PACKAGE
}
\note{
Development of this package was kindly supported by 
FUNDER under THIS GRANT NUMBER
}
\section{FDAopenR classess and functions}{

The main classes are:
\tabular{ll}{
  \code{FDA}         \tab Parent class \cr
  \code{Drug}        \tab FDA tested Drug class \cr
  \code{Device}      \tab Devices class \cr
  \code{Food}        \tab Food class \cr
  \code{Other}       \tab Other non Food, Drug, and Device classes}
The main methods are:
\tabular{ll}{
  \code{listSF}          \tab List Search Field for Each Classes \cr
  \code{extractFDA}      \tab Extract FDA data \cr
  \code{ndcToString}     \tab Convert numeric NDC to appropriate string with hyphens}
}

\references{
Kim et al. (2019). \emph{R package for  accessing and manipulating
openFDA API}. In preparation.
}
\seealso{
\href{https://open.fda.gov/apis/}{openFDA API}
}
