% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_drug.R
\name{find_drug}
\alias{find_drug}
\title{find_ndc Function: Find ndc from FDA ndc database}
\usage{
find_drug(src = "fda", ndc = NULL, bname = NULL, gname = NULL,
  dea_schedule = NULL, limit = NULL, query = NULL, ...)
}
\arguments{
\item{src}{String describing the source (e.g. \code{"fda"}).}

\item{ndc}{Integer describing the NDC number of drug of interest
(e.g. \code{"fda"})}

\item{bname}{String indicating brand name (e.g. \code{"lorazepam"})}

\item{gname}{String indicating generic name (e.g. \code{"lorazepam"})}

\item{dea_schedule}{Integer indicating class(e.g. \code{"1"})}

\item{limit}{Integer indicating limit on how many items to limit in
the result (e.g. \code{"1"})}

\item{query}{String indicating actual \code{httr::GET()} query to the
API}

\item{...}{Additional parameters specified (TBD)}
}
\value{
A list with components based on the parameters specified.
}
\description{
\code{find_ndc} returns a list of items after connecting to the openFDA API.
}
\examples{
find_drug(src="fda", ndc=687889736)
find_drug(src="fda", bname="Acetaminophen")
find_drug(src="fda", gname="Codeine")

\dontrun{
  find_drug(ndc="687889736")
}
}
\seealso{
\url{https://open.fda.gov/apis/drug/ndc/} for the actual
API description from openFDA. \code{\link{httr}} for accessing API.
}
