% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ndc.R
\name{find_ndc}
\alias{find_ndc}
\title{find_ndc Function: Find ndc from FDA ndc database}
\usage{
find_ndc(src = "fda", ndc = NULL, bname = NULL, gname = NULL,
  dea_schedule = NULL, limit = NULL, query = NULL, ...)
}
\arguments{
\item{src}{String describing the source (e.g. \code{"fda"})}

\item{ndc}{Integer describing the NDC number of drug of interest
(e.g. \code{"fda"})}

\item{bname}{String indicating brand name (e.g. \code{"lorazepam"})}

\item{gname}{String indicating generic name (e.g. \code{"lorazepam"})}

\item{dea_schedule}{Integer indicating class(e.g. \code{"1"})}

\item{limit}{Integer indicating limit on how many items to limit in the result
(e.g. \code{"1"})}

\item{query}{String indicating actual \code{httr::GET()} query to the API}

\item{...}{Additional parameters specified (TBD)}
}
\value{
A list with components based on the parameters specified.
}
\description{
find_ndc Function: Find ndc from FDA ndc database
}
