% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Food.R
\docType{class}
\name{Food-class}
\alias{Food-class}
\title{An S4 class to represent Food objects}
\value{
An object of a class Food
}
\description{
\code{new("Food")} will instatiate an object of class Food which holds
attributes and methods for useful for accessing the \href{https://open.fda.gov/apis/}{openFDA API}
}
\section{Slots}{

\describe{
\item{\code{reTab}}{data.frame object of Recall Enforcement Reports}

\item{\code{aeTab}}{data.frame object of Adverse Events.}
}}

\examples{
drug1 <- new("Food", apikey="Sd8Jfad834FDS")

\dontrun{
  drugnorun <- new("Food")
}
}
\seealso{
\url{https://open.fda.gov/apis/drug/ndc/} for the actual
API description from openFDA. \code{\link{jsonlite}} for accessing API.
}
