% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listSF.R
\docType{methods}
\name{listSF}
\alias{listSF}
\alias{listSF,character,ANY-method}
\alias{listSF,FDA-method}
\alias{listSF,Other-method}
\title{An S4 method to list Search Field for Each child Object of FDA class}
\usage{
listSF(Obj, endpoint, ...)

\S4method{listSF}{FDA}(Obj, endpoint, ...)

\S4method{listSF}{Other}(Obj, endpoint, ...)
}
\arguments{
\item{Obj}{An Obj of class FDA or child thereof.}

\item{endpoint}{string indicating the endpoint for the specific class.
For example, for the class \code{"Drug"}, the endpoint may be \code{"event"}.}

\item{\dots}{Passed down to searchlist within the function}
}
\value{
An object of a class Food
}
\description{
\code{listSF(Obj} will search for search fields appropriate for the 
instantiated \code{Obj}.Users can use the search field outputted in 
\code{extractFDA()} to extract relevant information
}
\examples{
drug1 <- new("Drug", apikey="Y3CgaZj67AotB7b4XLhzHJTY7oBWKUC3u1fYulw8")
drug1sf <- listSF(drug1, "event")

\dontrun{
  drug1sf <- listSF("Food")
  drug1sf <- listSF(drug1,"random") # there is no access point
}
}
\seealso{
\url{https://open.fda.gov/apis/drug/ndc/} for the actual
API description from openFDA. \code{\link{jsonlite}} for accessing API.
}
