% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndcToString.R
\docType{methods}
\name{ndcToString}
\alias{ndcToString}
\alias{ndcToString,character,ANY-method}
\alias{ndcToString,Drug-method}
\alias{ndcToString,Other-method}
\title{An S4 method to Convert numeric NDC values to string values for Drug and Other class}
\usage{
ndcToString(Obj)

\S4method{ndcToString}{Drug}(Obj)

\S4method{ndcToString}{Other}(Obj)
}
\arguments{
\item{Obj}{An Obj of class Drug or Other}
}
\value{
An object of a class Food
}
\description{
\code{ndcToString(Obj} will use the \code{Obj@ndc} slot to convert 
to appropriate ndc string for extracting information
}
\examples{
drug1 <- new("Food", apikey="Y3CgaZj67AotB7b4XLhzHJTY7oBWKUC3u1fYulw8")

\dontrun{
  drugnorun <- new("Food")
}
}
\seealso{
\url{https://open.fda.gov/apis/drug/ndc/} for the actual
API description from openFDA. \code{\link{jsonlite}} for accessing API.
}
