% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Other.R
\docType{class}
\name{Other-class}
\alias{Other-class}
\title{An S4 class to represent Other objects}
\value{
An object of a class Other
}
\description{
\code{new("Other")} will instatiate an object of class Other which holds
attributes and methods for useful for accessing the \href{https://open.fda.gov/apis/}{openFDA API}
}
\section{Slots}{

\describe{
\item{\code{nsdeTab}}{data.frame object of NDC SPL Data Elements File (NSDE)}

\item{\code{ndc}}{Numeric ndc value. This will be converted later using the
\code{\link{ndcToString}} method into \code{Obj@ndcquery} value.}

\item{\code{ndcquery}}{String of ndc value converted using \code{\link{ndcToString}}.}
}}

\examples{
drug1 <- new("Other", apikey="Sd8Jfad834FDS")

\dontrun{
  drugnorun <- new("Other")
}
}
\seealso{
\url{https://open.fda.gov/apis/drug/ndc/} for the actual
API description from openFDA. \code{\link{jsonlite}} for accessing API.
}
